+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * A new command `pubreg audit-log` for viewing audit logs of Maskinporten
   clients has been added.
 
+* Automatic client key rotation is now offered when creating a new client key.
+
 ## 1.2.0
 
 * The `pubreg delete-client` command now offers to delete the SSM key parameters